LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/responsive-iframe/responsive-iframe-value.tentative_value=content-block-size,400,150,vertical-rl-expected.txt


FAIL Measure target with value=content-block-size and wm=vertical-rl assert_array_equals: expected property 0 to be 400 but got 300 (expected array [400, 150] got [300, 150])
